Rename lib/confs to conf
