+++ scripts/kstest-runner
@@ -26,7 +26,7 @@ import sys
 import argparse
 
 from launcher.lib.virtual_controller import VirtualManager
-from launcher.lib.confs.configuration import VirtualConfiguration
+from launcher.lib.conf.configuration import VirtualConfiguration
 
 
 def create_parser():
